requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-config-url==0.1.1
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
Markdown==3.3.4
psycopg2-binary==2.9.1
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
Unipath==1.1
urllib3==1.26.7
